conda_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
appnope=0.1.2=py310hca03da5_1001
area=1.1.1=pypi_0
asttokens=2.0.5=pyhd3eb1b0_0
attrs=23.1.0=pypi_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=openblas
blosc=1.21.3=h313beb8_0
boost-cpp=1.82.0=h48ca7d4_2
bottleneck=1.3.5=py310h96f19d2_0
branca=0.6.0=py310hca03da5_0
brotli=1.0.9=h1a28f6b_7
brotli-bin=1.0.9=h1a28f6b_7
brotli-python=1.0.9=py310hc377ac9_7
bzip2=1.0.8=h620ffc9_4
c-ares=1.19.1=h80987f9_0
ca-certificates=2023.08.22=hca03da5_0
cairo=1.16.0=h302bd0f_5
certifi=2023.7.22=py310hca03da5_0
cffi=1.15.1=py310h80987f9_3
cfitsio=3.470=h7f6438f_7
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=pypi_0
click-plugins=1.1.1=pypi_0
cligj=0.7.2=pypi_0
comm=0.1.2=py310hca03da5_0
contourpy=1.0.5=py310h525c30c_0
cryptography=41.0.3=py310hd4332d6_0
cycler=0.11.0=pyhd3eb1b0_0
debugpy=1.6.7=py310h313beb8_0
decorator=5.1.1=pyhd3eb1b0_0
exceptiongroup=1.0.4=py310hca03da5_0
executing=0.8.3=pyhd3eb1b0_0
expat=2.5.0=h313beb8_0
ffmpeg=4.2.2=h04105a8_0
filelock=3.9.0=py310hca03da5_0
fiona=1.9.5=pypi_0
folium=0.14.0=py310hca03da5_0
fontconfig=2.14.1=hee714a5_2
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h1192e45_0
freexl=1.0.6=h1a28f6b_0
gdal=3.7.2=pypi_0
geographiclib=2.0=pypi_0
geopandas=0.14.0=pypi_0
geopy=2.4.0=pypi_0
georaster=1.25=pypi_0
geos=3.9.1=hc377ac9_1
geotiff=1.7.0=h41f0982_1
gettext=0.21.0=h13f89a0_1
giflib=5.2.1=h80987f9_3
glib=2.69.1=h514c7bf_2
gmp=6.2.1=hc377ac9_3
gmpy2=2.1.2=py310h8c48613_0
gnutls=3.6.15=h887c41c_0
hdf4=4.2.13=h5e329fb_3
hdf5=1.12.1=h05c076b_3
icu=73.1=h313beb8_0
idna=3.4=py310hca03da5_0
ipykernel=6.25.0=py310h33ce5c2_0
ipython=8.15.0=py310hca03da5_0
jedi=0.18.1=py310hca03da5_1
jinja2=3.1.2=py310hca03da5_0
jpeg=9e=h80987f9_1
json-c=0.16=h1a28f6b_0
jupyter_client=8.5.0=py310hca03da5_0
jupyter_core=5.5.0=py310hca03da5_0
kealib=1.5.0=hba2eb73_1
kiwisolver=1.4.4=py310h313beb8_0
krb5=1.20.1=hf3e1bf2_1
lame=3.100=h1a28f6b_0
lcms2=2.12=hba8e193_0
lerc=3.0=hc377ac9_0
libboost=1.82.0=h0bc93f9_2
libbrotlicommon=1.0.9=h1a28f6b_7
libbrotlidec=1.0.9=h1a28f6b_7
libbrotlienc=1.0.9=h1a28f6b_7
libcurl=8.4.0=h3e2b118_0
libcxx=14.0.6=h848a8c0_0
libdeflate=1.17=h80987f9_1
libedit=3.1.20221030=h80987f9_0
libev=4.33=h1a28f6b_1
libffi=3.4.4=hca03da5_0
libgdal=3.6.2=hdbe3b94_3
libgfortran=5.0.0=11_3_0_hca03da5_28
libgfortran5=11.3.0=h009349e_28
libiconv=1.16=h1a28f6b_2
libidn2=2.3.4=h80987f9_0
libjpeg-turbo=2.0.0=h1a28f6b_0
libkml=1.3.0=hc4d7c42_7
libllvm14=14.0.6=h7ec7a93_3
libnetcdf=4.8.1=h0fce390_4
libnghttp2=1.57.0=h62f6fdd_0
libopenblas=0.3.21=h269037a_0
libopus=1.3=h1a28f6b_1
libpng=1.6.39=h80987f9_0
libpq=12.15=h02f6b3c_1
libsodium=1.0.18=h1a28f6b_0
libspatialite=4.3.0a=h4707d77_23
libssh2=1.10.0=h02f6b3c_2
libtasn1=4.19.0=h80987f9_0
libtiff=4.5.1=h313beb8_0
libunistring=0.9.10=h1a28f6b_0
libvpx=1.10.0=hc377ac9_0
libwebp=1.3.2=ha3663a8_0
libwebp-base=1.3.2=h80987f9_0
libxml2=2.10.4=h0dcf63f_1
libzip=1.8.0=h62fee54_1
llvm-openmp=14.0.6=hc6e5704_0
llvmlite=0.41.0=py310h514c7bf_0
lz4-c=1.9.4=h313beb8_0
markupsafe=2.1.1=py310h1a28f6b_0
matplotlib-base=3.8.0=py310h46d7db6_0
matplotlib-inline=0.1.6=py310hca03da5_0
mpc=1.1.0=h8c48613_1
mpfr=4.0.2=h695f6f0_1
mpmath=1.3.0=py310hca03da5_0
munkres=1.1.4=py_0
ncurses=6.4=h313beb8_0
nest-asyncio=1.5.6=py310hca03da5_0
nettle=3.7.3=h84b5d62_1
networkx=3.2.1=pypi_0
nspr=4.35=h313beb8_0
nss=3.89.1=h313beb8_0
numba=0.58.0=py310h46d7db6_0
numexpr=2.8.7=py310hecc3335_0
numpy=1.24.3=py310hb93e574_0
numpy-base=1.24.3=py310haf87e8b_0
openh264=1.8.0=h98b2900_0
openjpeg=2.3.0=h7a6adac_2
openssl=3.0.11=h1a28f6b_2
osmnx=1.7.1=pypi_0
packaging=23.1=py310hca03da5_0
pandas=2.1.1=py310h46d7db6_0
parso=0.8.3=pyhd3eb1b0_0
pcre=8.45=hc377ac9_0
pcre2=10.42=hb066dcc_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=10.0.1=py310h3b245a6_0
pip=23.3=py310hca03da5_0
pixman=0.40.0=h1a28f6b_0
platformdirs=3.10.0=py310hca03da5_0
poppler=22.12.0=h52f4003_3
poppler-data=0.4.11=hca03da5_1
proj=7.2.0=heac154c_1
prompt-toolkit=3.0.36=py310hca03da5_0
psutil=5.9.0=py310h1a28f6b_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.15.1=py310hca03da5_1
pymongo=3.12.0=py310hc377ac9_0
pyopenssl=23.2.0=py310hca03da5_0
pyparsing=3.0.9=py310hca03da5_0
pyproj=3.6.1=pypi_0
pyshp=2.1.3=pyhd3eb1b0_0
pysocks=1.7.1=py310hca03da5_0
python=3.10.13=hb885b13_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-tzdata=2023.3=pyhd3eb1b0_0
pytorch=2.1.0=py3.10_0
pytz=2023.3.post1=py310hca03da5_0
pyyaml=6.0.1=py310h80987f9_0
pyzmq=25.1.0=py310h313beb8_0
qhull=2020.2=h48ca7d4_2
readline=8.2=h1a28f6b_0
requests=2.31.0=py310hca03da5_0
scipy=1.11.3=py310h20cbe94_0
seaborn=0.12.2=py310hca03da5_0
setuptools=68.0.0=py310hca03da5_0
shapely=2.0.1=py310h696e30d_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.41.2=h80987f9_0
stack_data=0.2.0=pyhd3eb1b0_0
sympy=1.11.1=py310hca03da5_0
tbb=2021.8.0=h48ca7d4_0
tiledb=2.3.3=hb4a6b97_3
tk=8.6.12=hb8d0fd4_0
torchaudio=2.1.0=py310_cpu
torchvision=0.16.0=py310_cpu
tornado=6.3.3=py310h80987f9_0
traitlets=5.7.1=py310hca03da5_0
typing_extensions=4.7.1=py310hca03da5_0
tzdata=2023c=h04d1e81_0
uriparser=0.9.7=h80987f9_0
urllib3=1.26.18=py310hca03da5_0
wcwidth=0.2.5=pyhd3eb1b0_0
wheel=0.41.2=py310hca03da5_0
x264=1!152.20180806=h1a28f6b_0
xerces-c=3.2.4=h313beb8_1
xz=5.4.2=h80987f9_0
yaml=0.2.5=h1a28f6b_0
zeromq=4.3.4=hc377ac9_0
zlib=1.2.13=h5a0b063_0
zstd=1.5.5=hd90d995_0
